+++ Swiften/Base/Debug.h
@@ -1,10 +1,14 @@
 /*
- * Copyright (c) 2015 Isode Limited.
+ * Copyright (c) 2015-2016 Isode Limited.
  * All rights reserved.
  * See the COPYING file for more information.
  */
 
+#pragma once
+
 #include <iosfwd>
+
+#include <Swiften/Client/ClientSession.h>
 
 namespace Swift {
     class ClientError;
@@ -18,3 +22,5 @@ namespace boost {
 std::ostream& operator<<(std::ostream& os, const Swift::ClientError& error);
 
 std::ostream& operator<<(std::ostream& os, Swift::Element* ele);
+
+std::ostream& operator<<(std::ostream& os, Swift::ClientSession::State state);