+++ Swiften/Client/ClientOptions.h
@@ -30,21 +30,7 @@ namespace Swift {
             HTTPConnectProxy
         };
 
-        ClientOptions() :
-                useStreamCompression(true),
-                useTLS(UseTLSWhenAvailable),
-                allowPLAINWithoutTLS(false),
-                useStreamResumption(false),
-                forgetPassword(false),
-                useAcks(true),
-                singleSignOn(false),
-                manualHostname(""),
-                manualPort(-1),
-                proxyType(SystemConfiguredProxy),
-                manualProxyHostname(""),
-                manualProxyPort(-1),
-                boshHTTPConnectProxyAuthID(""),
-                boshHTTPConnectProxyAuthPassword("") {
+        ClientOptions() {
         }
 
         /**
@@ -52,14 +38,14 @@ namespace Swift {
          *
          * Default: true
          */
-        bool useStreamCompression;
+        bool useStreamCompression = true;
 
         /**
          * Sets whether TLS encryption should be used.
          *
          * Default: UseTLSWhenAvailable
          */
-        UseTLS useTLS;
+        UseTLS useTLS = UseTLSWhenAvailable;
 
         /**
          * Sets whether plaintext authentication is
@@ -67,14 +53,14 @@ namespace Swift {
          *
          * Default: false
          */
-        bool allowPLAINWithoutTLS;
+        bool allowPLAINWithoutTLS = false;
 
         /**
          * Use XEP-196 stream resumption when available.
          *
          * Default: false
          */
-        bool useStreamResumption;
+        bool useStreamResumption = false;
 
         /**
          * Forget the password once it's used.
@@ -84,69 +70,69 @@ namespace Swift {
          *
          * Default: false
          */
-        bool forgetPassword;
+        bool forgetPassword = false;
 
         /**
          * Use XEP-0198 acks in the stream when available.
          * Default: true
          */
-        bool useAcks;
+        bool useAcks = true;
 
         /**
          * Use Single Sign On.
          * Default: false
          */
-        bool singleSignOn;
+        bool singleSignOn = false;
 
         /**
          * The hostname to connect to.
          * Leave this empty for standard XMPP connection, based on the JID domain.
          */
-        std::string manualHostname;
+        std::string manualHostname = "";
 
         /**
          * The port to connect to.
          * Leave this to -1 to use the port discovered by SRV lookups, and 5222 as a
          * fallback.
          */
-        int manualPort;
+        int manualPort = -1;
 
         /**
          * The type of proxy to use for connecting to the XMPP
          * server.
          */
-        ProxyType proxyType;
+        ProxyType proxyType = SystemConfiguredProxy;
 
         /**
          * Override the system-configured proxy hostname.
          */
-        std::string manualProxyHostname;
+        std::string manualProxyHostname = "";
 
         /**
          * Override the system-configured proxy port.
          */
-        int manualProxyPort;
+        int manualProxyPort = -1;
 
         /**
          * If non-empty, use BOSH instead of direct TCP, with the given URL.
          * Default: empty (no BOSH)
          */
-        URL boshURL;
+        URL boshURL = URL();
 
         /**
          * If non-empty, BOSH connections will try to connect over this HTTP CONNECT
          * proxy instead of directly.
          * Default: empty (no proxy)
          */
-        URL boshHTTPConnectProxyURL;
+        URL boshHTTPConnectProxyURL = URL();
 
         /**
          * If this and matching Password are non-empty, BOSH connections over
          * HTTP CONNECT proxies will use these credentials for proxy access.
          * Default: empty (no authentication needed by the proxy)
          */
-        SafeString boshHTTPConnectProxyAuthID;
-        SafeString boshHTTPConnectProxyAuthPassword;
+        SafeString boshHTTPConnectProxyAuthID = SafeString("");
+        SafeString boshHTTPConnectProxyAuthPassword = SafeString("");
 
         /**
          * This can be initialized with a custom HTTPTrafficFilter, which allows HTTP CONNECT
@@ -158,5 +144,11 @@ namespace Swift {
          * Options passed to the TLS stack
          */
         TLSOptions tlsOptions;
+
+        /**
+         * Session shutdown timeout in milliseconds. This is the maximum time Swiften
+         * waits from a session close to the socket close.
+         */
+        int sessionShutdownTimeoutInMilliseconds = 10000;
     };
 }