+++ Swiften/Client/ClientSession.h
@@ -16,18 +16,21 @@
 #include <Swiften/Elements/ToplevelElement.h>
 #include <Swiften/JID/JID.h>
 #include <Swiften/Session/SessionStream.h>
-#include <Swiften/StreamManagement/StanzaAckRequester.h>
-#include <Swiften/StreamManagement/StanzaAckResponder.h>
 
 namespace Swift {
-    class ClientAuthenticator;
     class CertificateTrustChecker;
-    class IDNConverter;
+    class ClientAuthenticator;
     class CryptoProvider;
+    class IDNConverter;
+    class Stanza;
+    class StanzaAckRequester;
+    class StanzaAckResponder;
+    class TimerFactory;
+    class Timer;
 
     class SWIFTEN_API ClientSession : public std::enable_shared_from_this<ClientSession> {
         public:
-            enum State {
+            enum class State {
                 Initial,
                 WaitingForStreamStart,
                 Negotiating,
@@ -55,7 +58,8 @@ namespace Swift {
                     SessionStartError,
                     TLSClientCertificateError,
                     TLSError,
-                    StreamError
+                    StreamError,
+                    StreamEndError, // The server send a closing stream tag.
                 } type;
                 std::shared_ptr<boost::system::error_code> errorCode;
                 Error(Type type) : type(type) {}
@@ -69,8 +73,8 @@ namespace Swift {
 
             ~ClientSession();
 
-            static std::shared_ptr<ClientSession> create(const JID& jid, std::shared_ptr<SessionStream> stream, IDNConverter* idnConverter, CryptoProvider* crypto) {
-                return std::shared_ptr<ClientSession>(new ClientSession(jid, stream, idnConverter, crypto));
+            static std::shared_ptr<ClientSession> create(const JID& jid, std::shared_ptr<SessionStream> stream, IDNConverter* idnConverter, CryptoProvider* crypto, TimerFactory* timerFactory) {
+                return std::shared_ptr<ClientSession>(new ClientSession(jid, stream, idnConverter, crypto, timerFactory));
             }
 
             State getState() const {
@@ -92,7 +96,6 @@ namespace Swift {
             void setUseAcks(bool b) {
                 useAcks = b;
             }
-
 
             bool getStreamManagementEnabled() const {
                 // Explicitly convert to bool. In C++11, it would be cleaner to
@@ -116,7 +119,7 @@ namespace Swift {
             void finish();
 
             bool isFinished() const {
-                return getState() == Finished;
+                return getState() == State::Finished;
             }
 
             void sendCredentials(const SafeByteArray& password);
@@ -138,6 +141,10 @@ namespace Swift {
                 authenticationPort = i;
             }
 
+            void setSessionShutdownTimeout(int timeoutInMilliseconds) {
+                sessionShutdownTimeoutInMilliseconds = timeoutInMilliseconds;
+            }
+
         public:
             boost::signals2::signal<void ()> onNeedCredentials;
             boost::signals2::signal<void ()> onInitialized;
@@ -150,7 +157,8 @@ namespace Swift {
                     const JID& jid,
                     std::shared_ptr<SessionStream>,
                     IDNConverter* idnConverter,
-                    CryptoProvider* crypto);
+                    CryptoProvider* crypto,
+                    TimerFactory* timerFactory);
 
             void finishSession(Error::Type error);
             void finishSession(std::shared_ptr<Swift::Error> error);
@@ -163,7 +171,9 @@ namespace Swift {
 
             void handleElement(std::shared_ptr<ToplevelElement>);
             void handleStreamStart(const ProtocolHeader&);
+            void handleStreamEnd();
             void handleStreamClosed(std::shared_ptr<Swift::Error>);
+            void handleStreamShutdownTimeout();
 
             void handleTLSEncrypted();
 
@@ -175,13 +185,17 @@ namespace Swift {
             void ack(unsigned int handledStanzasCount);
             void continueAfterTLSEncrypted();
             void checkTrustOrFinish(const std::vector<Certificate::ref>& certificateChain, std::shared_ptr<CertificateVerificationError> error);
+            void initiateShutdown(bool sendFooter);
 
         private:
             JID localJID;
             State state;
             std::shared_ptr<SessionStream> stream;
-            IDNConverter* idnConverter;
-            CryptoProvider* crypto;
+            IDNConverter* idnConverter = nullptr;
+            CryptoProvider* crypto = nullptr;
+            TimerFactory* timerFactory = nullptr;
+            std::shared_ptr<Timer> streamShutdownTimeout;
+            int sessionShutdownTimeoutInMilliseconds = 10000;
             bool allowPLAINOverNonTLS;
             bool useStreamCompression;
             UseTLS useTLS;